Rename generators submodule to topologies
