+++ dwave/graphs/topologies/chimera.py
