+++ dwave/graphs/topologies/common/__init__.py
@@ -12,4 +12,4 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-from dwave.graphs.generators.common.common import *
+from dwave.graphs.topologies.common.common import *
