+++ dwave/graphs/topologies/common/common.py
