+++ dwave/graphs/topologies/pegasus.py
